requirements.txt
cryptography
azure-functions
azure-functions-durable
azure-ai-documentintelligence
azure-storage-blob==12.25.1
azure-identity
azure-ai-formrecognizer
openai
requests
pandas
numpy
PyPDF2
